Fix: Error SonarCloud Code Analysis
