+++ src/app/components/info-row.tsx
@@ -1,0 +1,8 @@
+export function InfoRow({ label, value }: Readonly<{ label: string; value: string }>) {
+    return (
+        <div className="flex flex-col gap-0.5 sm:flex-row sm:gap-2">
+            <span className="min-w-36 text-sm font-medium text-foreground">{label}</span>
+            <span className="text-sm text-muted-foreground">{value}</span>
+        </div>
+    );
+}
